longened every double

+++ src/Cell.cpp
@@ -24,7 +24,7 @@
 #include <math.h>
 #include <vector>
 
-Cell::Cell(double x, double y, double width, double imagewidth, std::pair<double, double> *realMin, std::pair<double, double> *imageMin) : x(x), y(y), width(width), imagewidth(imagewidth) {
+Cell::Cell(long double x, long double y, long double width, long double imagewidth, std::pair<long double, long double> *realMin, std::pair<long double, long double> *imageMin) : x(x), y(y), width(width), imagewidth(imagewidth) {
 	imagex = abs(((realMin->first - x) / width ) * imagewidth) + imageMin->first;
     imagey = abs(((realMin->second - y) / width ) * imagewidth) + imageMin->second;
 
@@ -33,7 +33,7 @@ Cell::Cell(double x, double y, double width, double imagewidth, std::pair<double
 	complex = ComplexNumber(x, y);
 }
 
-void Cell::escape(ComplexNumber *c, std::vector<std::vector<Cell*>> *cells, std::pair<double, double> *realMin, unsigned int iterations, unsigned int cellsPerRow, unsigned int *maxCount, bool anti) {
+void Cell::escape(ComplexNumber *c, std::vector<std::vector<Cell*>> *cells, std::pair<long double, long double> *realMin, unsigned int iterations, unsigned int cellsPerRow, unsigned int *maxCount, bool anti) {
 	ComplexNumber z = ComplexNumber();
 	std::vector<Cell*> visited;
 	visited.clear();
@@ -58,8 +58,8 @@ void Cell::escape(ComplexNumber *c, std::vector<std::vector<Cell*>> *cells, std:
 }
 
 void Cell::render(unsigned int maxCount, unsigned int colourR, unsigned int colourG, unsigned int colourB) {
-    double percentageOfMax = log(counter) / log(maxCount);
-    double brightness = percentageOfMax > 0.25 ? percentageOfMax : 0.0;
+    long double percentageOfMax = log(counter) / log(maxCount);
+    long double brightness = percentageOfMax > 0.25 ? percentageOfMax : 0.0;
     glColor4f(colourR / 255.0f , colourG / 255.0f, colourB / 255.0f, brightness);
     glRectd(imagey, imagex, imagey + imagewidth, imagex + imagewidth); // x and y flipped to render it vertically
 }
